refactor(onboarding): move NewUserTutorial into core-ui and split TutorialItem
The You tab needs the same checklist card the wallet already draws, and a
feature module cannot reach another feature's internal package. core-ui is
where shared composables live and the convention plugin already puts it on
every feature module, so the move only changes the package. The strings and
drawables the component reads stay in :apps:flipcash:core, which core-ui
depends on, so the R import is unchanged.

TutorialItem gains nested Wallet and Profile families and the composable gains
a type parameter, so each screen's when stays exhaustive over the items it can
actually be given. The card radius moves to shapes.extraSmall, which is the
6 dp the design asks for.
